# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Autenticação para Teste Técnico Merchion",
        "description": "Documentação da API para registro e login de usuários.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor da API"
        }
    ],
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Registra um novo usuário",
                "operationId": "a718f172ff3ac464d723835815f8fb57",
                "requestBody": {
                    "description": "Dados do novo usuário",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Seu nome lindo"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "seu_melhor_email@email.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "uma_senha_forte"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuário registrado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Usuário registrado com sucesso!"
                                        },
                                        "user": {
                                            "$ref": "#/components/schemas/UserResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Realiza o login do usuário e inicia uma sessão",
                "operationId": "44212a9096e4b09358281e9ec8a0701d",
                "requestBody": {
                    "description": "Credenciais do usuário",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "seu_melhor_email@email.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "uma_senha_forte"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login bem-sucedido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Login realizado com sucesso."
                                        },
                                        "user": {
                                            "$ref": "#/components/schemas/UserResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Credenciais inválidas"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "UserResource": {
                "title": "UserResource",
                "description": "Recurso do Usuário",
                "properties": {
                    "id": {
                        "description": "ID do usuário",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Nome do usuário",
                        "type": "string",
                        "example": "Seu nome lindo"
                    },
                    "email": {
                        "description": "Email do usuário",
                        "type": "string",
                        "format": "email",
                        "example": "seu_melhor_email@email.com"
                    },
                    "joined_at": {
                        "description": "Data de cadastro do usuário",
                        "type": "string",
                        "example": "31/07/2025"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Autenticação",
            "description": "Autenticação"
        }
    ]
}